+++ tests/iter_test.py
@@ -1,54 +1,53 @@
 import unittest
 import funct
 
-import monad_std.iter.iter
+from monad_std.prelude import *
 from monad_std.iter import IterMeta
-from monad_std import Option, Result, Ok, Err
 
 
 class ResultTest(unittest.TestCase):
     def test_iter(self):
-        t = IterMeta.iter(range(10))
+        t = siter(range(10))
         self.assertListEqual(t.collect_list(), list(range(10)))
-        t = IterMeta.iter(range(10))
+        t = siter(range(10))
         self.assertTupleEqual(t.collect_tuple(), tuple(range(10)))
-        t = IterMeta.iter(range(10))
+        t = siter(range(10))
         self.assertEqual(t.collect_string(), "".join(map(str, range(10))))
-        t = IterMeta.iter(range(10))
+        t = siter(range(10))
         self.assertEqual(t.collect_array(), funct.Array(range(10)))
-        t = IterMeta.iter(range(10))
+        t = siter(range(10))
         self.assertEqual(t.count(), 10)
 
         a = [1, 2, 3, 4]
-        it = IterMeta.iter(a)
+        it = siter(a)
         self.assertEqual(it.advance_by(2), Result.of_ok(None))
         self.assertEqual(it.next(), Option.some(3))
         self.assertEqual(it.advance_by(0), Result.of_ok(None))
         self.assertEqual(it.advance_by(100), Result.of_err(99))
 
         a = [1, 2, 3]
-        self.assertEqual(IterMeta.iter(a).nth(1), Option.some(2))
+        self.assertEqual(siter(a).nth(1), Option.some(2))
         a = [1, 2, 3]
-        self.assertEqual(IterMeta.iter(a).nth(10), Option.none())
+        self.assertEqual(siter(a).nth(10), Option.none())
 
         a = [1, 2, 3]
-        self.assertEqual(IterMeta.iter(a).last(), Option.some(3))
+        self.assertEqual(siter(a).last(), Option.some(3))
         a = []
-        self.assertEqual(IterMeta.iter(a).last(), Option.none())
+        self.assertEqual(siter(a).last(), Option.none())
 
         a = [1, 2, 3]
-        it = IterMeta.iter(a)
+        it = siter(a)
         self.assertEqual(it.next_chunk(2), Ok([1, 2]))
         self.assertEqual(it.next_chunk(2), Err([3]))
         quote = "not all those who wander are lost"
-        first, second, third = IterMeta.iter(quote.split(' ')).next_chunk(3).unwrap()
+        first, second, third = siter(quote.split(' ')).next_chunk(3).unwrap()
         self.assertEqual(first, 'not')
         self.assertEqual(second, 'all')
         self.assertEqual(third, 'those')
 
     def test_iter_enumerator(self):
         a = ["a", "b", "c"]
-        it = IterMeta.iter(a).enumerate()
+        it = siter(a).enumerate()
         self.assertEqual(it.next(), Option.some((0, "a")))
         self.assertEqual(it.next(), Option.some((1, "b")))
         self.assertEqual(it.next(), Option.some((2, "c")))
@@ -56,14 +55,14 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_filter(self):
         a = [-1, 0, 1, 2]
-        it = IterMeta.iter(a)
+        it = siter(a)
         self.assertListEqual(it.filter(lambda x: x > 0).collect_list(), [1, 2])
 
     def test_iter_filter_map(self):
         a = ["1", "two", "3.0", "four", "5"]
-        it1 = IterMeta.iter(a).filter_map(lambda x: Result.catch_from(float, x).ok())
+        it1 = siter(a).filter_map(lambda x: Result.catch_from(float, x).ok())
         it2 = (
-            IterMeta.iter(a)
+            siter(a)
             .map(lambda x: Result.catch_from(float, x))
             .filter(lambda x: x.is_ok())
             .map(lambda x: x.unwrap())
@@ -73,19 +72,19 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_flatten(self):
         a = [[1, 2, 3, 4], [5, 6]]
-        ftd = IterMeta.iter(a).flatten().collect_list()
+        ftd = siter(a).flatten().collect_list()
         self.assertListEqual(ftd, [1, 2, 3, 4, 5, 6])
 
         words = ["alpha", "beta", "gamma"]
-        ftd = IterMeta.iter(words).map(iter).flatten().collect_string()
+        ftd = siter(words).map(iter).flatten().collect_string()
         self.assertEqual(ftd, "alphabetagamma")
 
         a = [Option.some(123), Result.of_ok(321), Option.none(), Option.some(233), Result.of_err("err")]
-        ftd = IterMeta.iter(a).flatten().collect_list()
+        ftd = siter(a).flatten().collect_list()
         self.assertListEqual(ftd, [123, 321, 233])
 
         words = ["alpha", "beta", "gamma"]
-        merged = IterMeta.iter(words).flat_map(iter).collect_string()
+        merged = siter(words).flat_map(iter).collect_string()
         self.assertEqual(merged, "alphabetagamma")
 
     def test_iter_fuse(self):
@@ -115,7 +114,7 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_inspect(self):
         a = [1, 4, 2, 3]
-        sumed = (IterMeta.iter(a)
+        sumed = (siter(a)
                .inspect(lambda x: None)
                .filter(lambda x: x % 2 == 0)
                .inspect(lambda x: None)
@@ -123,7 +122,7 @@ class ResultTest(unittest.TestCase):
         self.assertEqual(sumed, 6)
 
     def test_iter_intersperse(self):
-        it = IterMeta.iter([0, 1, 2]).intersperse(100)
+        it = siter([0, 1, 2]).intersperse(100)
         self.assertEqual(it.next(), Option.some(0))
         self.assertEqual(it.next(), Option.some(100))
         self.assertEqual(it.next(), Option.some(1))
@@ -131,30 +130,30 @@ class ResultTest(unittest.TestCase):
         self.assertEqual(it.next(), Option.some(2))
         self.assertEqual(it.next(), Option.none())
 
-        hello = IterMeta.iter(["Hello", "World", "!"]).intersperse(' ').collect_string()
+        hello = siter(["Hello", "World", "!"]).intersperse(' ').collect_string()
         self.assertEqual(hello, "Hello World !")
 
-        src = IterMeta.iter(["Hello", "to", "all", "people", "!!"])
-        happy_emojis = IterMeta.iter([" ❤️ ", " 😀 "])
+        src = siter(["Hello", "to", "all", "people", "!!"])
+        happy_emojis = siter([" ❤️ ", " 😀 "])
         separator = lambda: happy_emojis.next().unwrap_or(" 🦀 ")
         result = src.intersperse_with(separator).collect_string()
         self.assertEqual(result, "Hello ❤️ to 😀 all 🦀 people 🦀 !!")
 
     # noinspection PyMethodMayBeStatic
     def test_iter_for_each(self):
-        (IterMeta.iter(range(5)).flat_map(lambda x: range(x * 100, x * 110))
+        (siter(range(5)).flat_map(lambda x: range(x * 100, x * 110))
          .enumerate()
          .filter(lambda d: (d[0] + d[1]) % 3 == 0)
          .for_each(lambda d: None))
 
     def test_iter_map(self):
-        t = IterMeta.iter(range(10))
+        t = siter(range(10))
         self.assertListEqual(
             t.map(lambda x: x * 2).filter(lambda x: x % 3 == 1).enumerate().collect_list(),
             list(enumerate(filter(lambda x: x % 3 == 1, map(lambda x: x * 2, range(10))))),
         )
         a = [1, 2, 3]
-        it = IterMeta.iter(a).map(lambda x: x * 2)
+        it = siter(a).map(lambda x: x * 2)
         self.assertEqual(it.next(), Option.some(2))
         self.assertEqual(it.next(), Option.some(4))
         self.assertEqual(it.next(), Option.some(6))
@@ -162,7 +161,7 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_peek(self):
         xs = [1, 2, 3]
-        it = IterMeta.iter(xs).peekable()
+        it = siter(xs).peekable()
         self.assertEqual(it.peek(), Option.some(1))
         self.assertEqual(it.next(), Option.some(1))
         self.assertEqual(it.peek(), Option.some(2))
@@ -174,43 +173,43 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_fold(self):
         a = [1, 2, 3]
-        it = IterMeta.iter(a)
+        it = siter(a)
         self.assertEqual(it.fold(0, lambda acc, x: acc + x), 6)
 
         numbers = [1, 2, 3, 4, 5]
-        result = IterMeta.iter(numbers).fold("0", lambda acc, x: f"({acc} + {x})")
+        result = siter(numbers).fold("0", lambda acc, x: f"({acc} + {x})")
         self.assertEqual(result, "(((((0 + 1) + 2) + 3) + 4) + 5)")
 
     def test_iter_index(self):
         a = [1, 2, 3]
-        self.assertEqual(IterMeta.iter(a).find(lambda x: x == 2), Option.some(2))
-        self.assertEqual(IterMeta.iter(a).find(lambda x: x == 5), Option.none())
+        self.assertEqual(siter(a).find(lambda x: x == 2), Option.some(2))
+        self.assertEqual(siter(a).find(lambda x: x == 5), Option.none())
 
         a = ["lol", "wow", "2", "5"]
-        res = IterMeta.iter(a).find_map(lambda x: Result.catch_from(int, x).ok())
+        res = siter(a).find_map(lambda x: Result.catch_from(int, x).ok())
         self.assertEqual(res, Option.some(2))
 
         a = [1, 2, 3]
 
-        self.assertTrue(IterMeta.iter(a).exist(2))
-        self.assertFalse(IterMeta.iter(a).exist(5))
+        self.assertTrue(siter(a).exist(2))
+        self.assertFalse(siter(a).exist(5))
 
-        self.assertEqual(IterMeta.iter(a).position(lambda x: x == 2), Option.some(1))
-        self.assertEqual(IterMeta.iter(a).position(lambda x: x == 5), Option.none())
+        self.assertEqual(siter(a).position(lambda x: x == 2), Option.some(1))
+        self.assertEqual(siter(a).position(lambda x: x == 5), Option.none())
 
-        self.assertEqual(IterMeta.iter(a).index(2), Option.some(1))
-        self.assertEqual(IterMeta.iter(a).index(5), Option.none())
+        self.assertEqual(siter(a).index(2), Option.some(1))
+        self.assertEqual(siter(a).index(5), Option.none())
 
     def test_iter_reduce(self):
-        reduced = IterMeta.iter(range(10)).reduce(lambda acc, e: acc + e)
+        reduced = siter(range(10)).reduce(lambda acc, e: acc + e)
         self.assertEqual(reduced, Option.some(45))
-        self.assertEqual(reduced.unwrap(), IterMeta.iter(range(10)).fold(0, lambda acc, e: acc + e))
+        self.assertEqual(reduced.unwrap(), siter(range(10)).fold(0, lambda acc, e: acc + e))
 
         a = [1, 2, 3]
-        self.assertEqual(IterMeta.iter(a).sum(), Option.some(6))
+        self.assertEqual(siter(a).sum(), Option.some(6))
 
-        self.assertEqual(IterMeta.iter(range(1, 6)).product(), Option.some(120))
-        self.assertEqual(IterMeta.iter(range(1, 1)).product(), Option.none())
+        self.assertEqual(siter(range(1, 6)).product(), Option.some(120))
+        self.assertEqual(siter(range(1, 1)).product(), Option.none())
 
     def test_iter_scan(self):
         a = [1, 2, 3, 4]
@@ -223,7 +222,7 @@ class ResultTest(unittest.TestCase):
                 res = Option.some(-st)
             return st, res
 
-        it = IterMeta.iter(a).scan(1, scanner)
+        it = siter(a).scan(1, scanner)
 
         self.assertEqual(it.next(), Option.some(-1))
         self.assertEqual(it.next(), Option.some(-2))
@@ -232,14 +231,14 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_skip(self):
         a = [1, 2, 3]
-        it = IterMeta.iter(a).skip(2)
+        it = siter(a).skip(2)
         self.assertEqual(it.next(), Option.some(3))
         self.assertEqual(it.next(), Option.none())
         self.assertEqual(it.next(), Option.none())
 
     def test_iter_take(self):
         a = [1, 2, 3]
-        it = IterMeta.iter(a).take(2)
+        it = siter(a).take(2)
         self.assertEqual(it.next(), Option.some(1))
         self.assertEqual(it.next(), Option.some(2))
         self.assertEqual(it.next(), Option.none())
@@ -255,25 +254,25 @@ class ResultTest(unittest.TestCase):
                 _f1 = _f2
                 _f2 = _f3
 
-        it = IterMeta.iter(fib()).take(4)
+        it = siter(fib()).take(4)
         self.assertListEqual(it.collect_list(), [1, 1, 2, 3])
-        it = IterMeta.iter(fib()).skip(5).take(5)
+        it = siter(fib()).skip(5).take(5)
         self.assertListEqual(it.collect_list(), [8, 13, 21, 34, 55])
 
         a = [-1, 0, 1]
-        it = IterMeta.iter(a).take_while(lambda v: v < 0)
+        it = siter(a).take_while(lambda v: v < 0)
         self.assertEqual(it.next(), Option.some(-1))
         self.assertEqual(it.next(), Option.none())
 
         a = [-1, 0, 1, -2]
-        it = IterMeta.iter(a).take_while(lambda v: v < 0)
+        it = siter(a).take_while(lambda v: v < 0)
         self.assertEqual(it.next(), Option.some(-1))
         self.assertEqual(it.next(), Option.none())
         self.assertEqual(it.next(), Option.none())
         self.assertEqual(it.next(), Option.none())
 
         a = [1, 2, 3, 4]
-        it = IterMeta.iter(a)
+        it = siter(a)
         result = it.take_while(lambda v: v != 3).collect_list()
         self.assertListEqual(result, [1, 2])
         result = it.collect_list()
@@ -282,7 +281,7 @@ class ResultTest(unittest.TestCase):
     def test_iter_zip(self):
         a1 = [1, 3, 5]
         a2 = [2, 4, 6]
-        it = IterMeta.iter(a1).zip(IterMeta.iter(a2))
+        it = siter(a1).zip(siter(a2))
         self.assertEqual(it.next(), Option.some((1, 2)))
         self.assertEqual(it.next(), Option.some((3, 4)))
         self.assertEqual(it.next(), Option.some((5, 6)))
@@ -290,30 +289,72 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_chain(self):
         element = 1
-        it = IterMeta.once(element)
+        it = once(element)
         self.assertEqual(it.next(), Option.some(1))
         self.assertEqual(it.next(), Option.none())
 
         a1 = [1, 3, 5]
         a2 = [2, 4, 6]
-        it1 = IterMeta.iter(a1)
-        it2 = IterMeta.iter(a2)
+        it1 = siter(a1)
+        it2 = siter(a2)
         self.assertListEqual(it1.chain(it2).collect_list(), [1, 3, 5, 2, 4, 6])
 
+    def test_iter_once(self):
+        it = once(1)
+        self.assertEqual(it.next(), Option.some(1))
+        self.assertEqual(it.next(), Option.none())
+
+        flag = 0
+
+        def closure() -> int:
+            nonlocal flag
+            flag += 1
+            return flag
+
+        it = once_with(closure)
+        self.assertEqual(flag, 0)
+        self.assertEqual(it.next(), Option.some(1))
+        self.assertEqual(flag, 1)
+        self.assertEqual(it.next(), Option.none())
+        self.assertEqual(flag, 1)
+
+        it = once_with(lambda: 1)
+        self.assertEqual(it.advance_by(0), Ok(None))
+        self.assertEqual(it.advance_by(1), Ok(None))
+        self.assertEqual(it.advance_by(100), Err(100))
+        it = once_with(lambda: 1)
+        self.assertEqual(it.advance_by(100), Err(99))
+
+        it = once_with(lambda: 1)
+        self.assertEqual(it.next_chunk(1), Ok([1]))
+        self.assertEqual(it.next_chunk(1), Err([]))
+        it = once_with(lambda: 1)
+        self.assertEqual(it.next_chunk(2), Err([1]))
+
+        it = once_with(lambda: 1)
+        self.assertEqual(it.nth(0), Option.some(1))
+        self.assertEqual(it.nth(0), Option.none())
+        it = once_with(lambda: 1)
+        self.assertEqual(it.nth(1), Option.none())
+
+    def test_iter_repeat(self):
+        it = repeat(5)
+        self.assertListEqual(it.take(5).collect_list(), [5] * 5)
+
     def test_iter_chunk(self):
-        a = IterMeta.iter("loerm").array_chunk(2)
+        a = siter("loerm").array_chunk(2)
         self.assertEqual(a.next(), Option.some(["l", "o"]))
         self.assertEqual(a.next(), Option.some(["e", "r"]))
         self.assertEqual(a.next(), Option.none())
         self.assertListEqual(a.get_unused().unwrap(), ["m"])
         self.assertListEqual(a.get_unused().unwrap(), ["m"])
 
-        it = IterMeta.iter([1, 2, 3, 4]).array_chunk(3)
+        it = siter([1, 2, 3, 4]).array_chunk(3)
         self.assertEqual(it.next(), Option.some([1, 2, 3]))
         self.assertEqual(it.next(), Option.none())
         self.assertEqual(it.get_unused(), Option.some([4]))
 
-        a = IterMeta.iter("loerm").chunk(2)
+        a = siter("loerm").chunk(2)
         self.assertEqual(a.next(), Option.some(["l", "o"]))
         self.assertEqual(a.next(), Option.some(["e", "r"]))
         self.assertEqual(a.next(), Option.some(["m"]))
@@ -321,18 +362,18 @@ class ResultTest(unittest.TestCase):
 
     def test_iter_check(self):
         a = [1, 2, 3]
-        self.assertTrue(IterMeta.iter(a).all(lambda x: x > 0))
-        self.assertFalse(IterMeta.iter(a).all(lambda x: x > 2))
+        self.assertTrue(siter(a).all(lambda x: x > 0))
+        self.assertFalse(siter(a).all(lambda x: x > 2))
         a = [1, 2, 3]
-        it = IterMeta.iter(a)
+        it = siter(a)
         self.assertFalse(it.all(lambda x: x != 2))
         self.assertEqual(it.next(), Option.some(3))
 
         a = [1, 2, 3]
-        self.assertTrue(IterMeta.iter(a).any(lambda x: x > 2))
-        self.assertFalse(IterMeta.iter(a).any(lambda x: x > 5))
+        self.assertTrue(siter(a).any(lambda x: x > 2))
+        self.assertFalse(siter(a).any(lambda x: x > 5))
         a = [1, 2, 3]
-        it = IterMeta.iter(a)
+        it = siter(a)
         self.assertTrue(it.any(lambda x: x != 2))
         self.assertEqual(it.next(), Option.some(2))